More reorganization of the file system
