+++ js/Analytics/EventDashboard.js
@@ -13,7 +13,7 @@ import {
   where,
 } from "https://www.gstatic.com/firebasejs/12.3.0/firebase-firestore.js";
 import { onAuthStateChanged } from "https://www.gstatic.com/firebasejs/12.3.0/firebase-auth.js";
-import { auth, db } from "../../Shared/firebase-config.js";
+import { auth, db } from "../../js/Shared/firebase-config.js";
 
 const params = new URLSearchParams(window.location.search);
 const eventId = params.get("id"); // ✅ Single-event mode if ?id=
