+++ js/Events/LoadEvents.js
@@ -1,6 +1,6 @@
 // loadEvents.js — works for both Organizer and Student dashboards
 
-import { auth, db } from "../../Shared/firebase-config.js";
+import { auth, db } from "../../js/Shared/firebase-config.js";
 import {
   collection,
   getDocs,
