+++ js/Registration/auth.js
@@ -1,4 +1,4 @@
-import { auth, db } from "../../Shared/firebase-config.js";
+import { auth, db } from "../../js/Shared/firebase-config.js";
 
 import {
   createUserWithEmailAndPassword,
