+++ website/Administrator/admin-dashboard.html
@@ -139,10 +139,9 @@
     document.getElementById('clear-search').addEventListener('click', clearSearchView);
   </script>
 
-  <script type="module" src="../../Shared/firebase-config.js"></script>
   <script type="module" src="../../js/Registration/auth.js"></script>
   <script type="module" src="../../js/Events/LoadEvents.js"></script>
-  <script type="module" src="../../shared/scroll-controls.js"></script>
+  <script type="module" src="../../js/Shared/scroll-controls.js"></script>
 
 </body>
