+++ website/Organizer/EventDashboard.html
@@ -93,10 +93,9 @@
   </div>
 
   <!-- ✅ Correct Script Paths -->
-     <script type="module" src="../../Shared/firebase-config.js"></script>
-    <script type="module" src="../../js/Registration/auth.js"></script>
+  <script type="module" src="../../js/Registration/auth.js"></script>
   <script type="module" src="../../js/Analytics/EventDashboard.js"></script>
- 
+
 </body>
 
 </html>
