+++ website/Organizer/eventCreation.html
@@ -192,7 +192,6 @@
   </script>
 
   <!-- Firebase Scripts -->
-  <script type="module" src="../../Shared/firebase-config.js"></script>
   <script type="module" src="../../js/Registration/auth.js"></script>
   <script type="module" src="../../js/Events/events.js"></script>
 </body>
