+++ website/Organizer/eventEdit.html
@@ -212,7 +212,6 @@
   </script>
 
   <!-- Firebase + Auth -->
-  <script type="module" src="../../Shared/firebase-config.js"></script>
   <script type="module" src="../../js/Registration/auth.js"></script>
 
   <!-- Combined Events Script -->
