+++ website/Organizer/organizer-dashboard.html
@@ -129,11 +129,9 @@
   </script>
 
   <!-- Firebase + Auth + Events -->
-  <script type="module" src="../../Shared/firebase-config.js"></script>
-
   <script type="module" src="../../js/Registration/auth.js"></script>
   <script type="module" src="../../js/Events/LoadEvents.js"></script>
-  <script type="module" src="../../shared/scroll-controls.js"></script>
+  <script type="module" src="../../js/Shared/scroll-controls.js"></script>
 
 </body>
