+++ website/Student/eventPageStu.html
@@ -46,7 +46,7 @@
 
   <!-- Firebase scripts -->
   <script type="module">
-    import { auth, db } from "../../Shared/firebase-config.js";
+    import { auth, db } from "../../js/Shared/firebase-config.js";
     import { doc, getDoc, updateDoc, setDoc, arrayUnion, arrayRemove } from "https://www.gstatic.com/firebasejs/12.3.0/firebase-firestore.js";
     import { onAuthStateChanged } from "https://www.gstatic.com/firebasejs/12.3.0/firebase-auth.js";
