+++ website/Student/myTickets.html
@@ -37,12 +37,11 @@
   </div>
 
   <!-- Firebase + Auth -->
-  <script type="module" src="../../Shared/firebase-config.js"></script>
   <script type="module" src="../../js/Registration/auth.js"></script>
 
   <!-- 🟢 Load actual tickets -->
   <script type="module">
-    import { auth, db } from "../../Shared/firebase-config.js";
+    import { auth, db } from "../../js/Shared/firebase-config.js";
     import { onAuthStateChanged } from "https://www.gstatic.com/firebasejs/12.3.0/firebase-auth.js";
     import { doc, getDoc, collection, getDocs } from "https://www.gstatic.com/firebasejs/12.3.0/firebase-firestore.js";
